upgrade class space size from short to int
git-svn-id: http://bbs.cs.nthu.edu.tw/svn/maplebbs/trunk@306 bd1103fd-f9b1-db11-a66c-000d613fc1d9

+++ include/struct.h
@@ -418,7 +418,7 @@ typedef	struct
 {
   int count;
   char title[CH_TTLEN];
-  short chno[0];
+  int chno[0];
 }	ClassHeader;
