+++ maple/board.c
@@ -1042,7 +1042,7 @@ static void
 czh_load()
 {
   int fsize, chn, min_chn;
-  short *chx;
+  int *chx;
   char *img, fpath[64];
   CZH *chead, *ctail, *czh;
 
@@ -1060,7 +1060,7 @@ czh_load()
     {
       for (chn = CH_END - 2;; chn--)	/* 在所有的分類中找一個板名相同的 */
       {
-	chx = (short *) img + (CH_END - chn);
+	chx = (int *) img + (CH_END - chn);
 	if (!strncmp(czh->brdname, img + *chx, BNLEN))
 	{
 	  class_bits[-chn] |= BRD_Z_BIT;
@@ -1280,8 +1280,8 @@ static int class_hot = 0;
 
 static int
 mantime_cmp(a, b)	/* smiler.070602: 看板人氣排序 */
-short *a;
-short *b;
+int *a;
+int *b;
 {
   return bshm->mantime[*b] - bshm->mantime[*a];
 }
@@ -1294,7 +1294,7 @@ static int
 class_load(xo)
   XO *xo;
 {
-  short *cbase, *chead, *ctail;
+  int *cbase, *chead, *ctail;
   int chn;			/* ClassHeader number */
   int pos, max, val, zap;
   int bnum = 0;			/* smiler.070602: for熱門看板 */
@@ -1307,21 +1307,21 @@ class_load(xo)
 
   chn = CH_END - xo->key;
 
-  cbase = (short *) class_img;
+  cbase = (int *) class_img;
   chead = cbase + chn;
 
   pos = chead[0] + CH_TTLEN;
   max = chead[1];
 
-  chead = (short *) ((char *) cbase + pos);
-  ctail = (short *) ((char *) cbase + max);
+  chead = (int *) ((char *) cbase + pos);
+  ctail = (int *) ((char *) cbase + max);
 
   max -= pos;
 
-  if (cbase = (short *) xo->xyz)
-    cbase = (short *) realloc(cbase, max);
+  if (cbase = (int *) xo->xyz)
+    cbase = (int *) realloc(cbase, max);
   else
-    cbase = (short *) malloc(max);
+    cbase = (int *) malloc(max);
 
   xo->xyz = (char *) cbase;
 
@@ -1357,7 +1357,7 @@ class_load(xo)
   if (class_hot && bnum > 0)
   {
     cbase -= bnum;
-    xsort(cbase, bnum, sizeof(short), mantime_cmp);
+    xsort(cbase, bnum, sizeof(int), mantime_cmp);
   }
 
   xo->max = max;
@@ -1702,12 +1702,12 @@ class_bar(xo, mode)
   XO *xo;
   int mode;
 {
-  short *chp;
+  int *chp;
   BRD *brd;
   int chn, cnt, brdpost;
 
   cnt = xo->pos + 1;
-  chp = (short *) xo->xyz + xo->pos;
+  chp = (int *) xo->xyz + xo->pos;
   chn = *chp;
   brd = bshm->bcache + chn;
   brdpost = class_flag & UFO_BRDPOST;
@@ -1721,11 +1721,11 @@ class_bar(xo, mode)
   }
   else
   {
-    short *chx;
+    int *chx;
     char *img, *str;
 
     img = class_img;
-    chx = (short *) img + (CH_END - chn);
+    chx = (int *) img + (CH_END - chn);
     str = img + *chx;
     prints("%s%6d%c  %-13.13s\033[1;3%dm%5.5s\033[m%s%-*.*s%s",
       mode ? UCBAR[UCBAR_BRD] : "",
@@ -1746,7 +1746,7 @@ static int
 class_body(xo)
   XO *xo;
 {
-  short *chp;
+  int *chp;
   BRD *bcache;
   int n, cnt, max, chn, brdpost;
 #ifdef AUTO_JUMPBRD
@@ -1765,7 +1765,7 @@ class_body(xo)
   {
     class_jumpnext = 0;
     n = xo->pos;
-    chp = (short *) xo->xyz + n;
+    chp = (int *) xo->xyz + n;
 
     while (n < max)
     {
@@ -1798,7 +1798,7 @@ class_body(xo)
 #endif
 
   brdpost = class_flag & UFO_BRDPOST;
-  chp = (short *) xo->xyz + cnt;
+  chp = (int *) xo->xyz + cnt;
 
   n = 3;
   move(3, 0);
@@ -1818,11 +1818,11 @@ class_body(xo)
       }
       else			/* 分類群組 */
       {
-	short *chx;
+	int *chx;
 	char *img, *str;
 
 	img = class_img;
-	chx = (short *) img + (CH_END - chn);
+	chx = (int *) img + (CH_END - chn);
 	str = img + *chx;
 	prints("%6d%c  %-13.13s\033[1;3%dm%-5.5s\033[m%s\n",
 	  cnt, class_bits[-chn] & BRD_Z_BIT ? TOKEN_ZAP_BRD : ' ',
@@ -1930,7 +1930,7 @@ class_search(xo)
 
   if (vget(b_lines, 0, MSG_BID, buf, BNLEN + 1, DOECHO))
   {
-    short *chp, chn;
+    int *chp, chn;
     BRD *bcache, *brd;
 
     str_lowest(buf, buf);
@@ -1938,7 +1938,7 @@ class_search(xo)
     bcache = bshm->bcache;
     pos = num = xo->pos;
     max = xo->max;
-    chp = (short *) xo->xyz;
+    chp = (int *) xo->xyz;
 
     do
     {
@@ -1970,7 +1970,7 @@ class_searchBM(xo)
 
   if (vget(b_lines, 0, "請輸入板主：", buf, IDLEN + 1, DOECHO))
   {
-    short *chp, chn;
+    int *chp, chn;
     BRD *bcache, *brd;
 
     str_lower(buf, buf);
@@ -1978,7 +1978,7 @@ class_searchBM(xo)
     bcache = bshm->bcache;
     pos = num = xo->pos;
     max = xo->max;
-    chp = (short *) xo->xyz;
+    chp = (int *) xo->xyz;
 
     do
     {
@@ -2021,12 +2021,12 @@ class_zap(xo)
   XO *xo;
 {
   BRD *brd;
-  short *chp;
+  int *chp;
   int pos, num, chn;
   char token;
 
   pos = xo->pos;
-  chp = (short *) xo->xyz + pos;
+  chp = (int *) xo->xyz + pos;
   chn = *chp;
   if (chn >= 0)		/* 一般看板 */
   {
@@ -2052,7 +2052,7 @@ class_zap(xo)
   }
   else			/* 分類群組 */
   {
-    short *chx;
+    int *chx;
     char *img, brdname[BNLEN + 1];
 
     /* itoc.010909: 要隨 class_body() 版面變 */
@@ -2061,7 +2061,7 @@ class_zap(xo)
     outc(num & BRD_Z_BIT ? TOKEN_ZAP_BRD : ' ');
 
     img = class_img;
-    chx = (short *) img + (CH_END - chn);
+    chx = (int *) img + (CH_END - chn);
     str_ncpy(brdname, img + *chx, BNLEN + 1);
     czh_put(brdname);
   }
@@ -2109,10 +2109,10 @@ static int
 class_visit(xo)		/* itoc.010128: 看板列表設定看板已讀 */
   XO *xo;
 {
-  short *chp;
+  int *chp;
   int chn;
 
-  chp = (short *) xo->xyz + xo->pos;
+  chp = (int *) xo->xyz + xo->pos;
   chn = *chp;
   if (chn >= 0)
   {
@@ -2132,10 +2132,10 @@ static int
 class_unvisit(xo)		/* itoc.010129: 看板列表設定看板未讀 */
   XO *xo;
 {
-  short *chp;
+  int *chp;
   int chn;
 
-  chp = (short *) xo->xyz + xo->pos;
+  chp = (int *) xo->xyz + xo->pos;
   chn = *chp;
   if (chn >= 0)
   {
@@ -2156,13 +2156,13 @@ class_nextunread(xo)
   XO *xo;
 {
   int max, pos, chn;
-  short *chp;
+  int *chp;
   BRD *bcache, *brd;
 
   bcache = bshm->bcache;
   max = xo->max;
   pos = xo->pos;
-  chp = (short *) xo->xyz + pos;
+  chp = (int *) xo->xyz + pos;
 
   while (++pos < max)
   {
@@ -2192,10 +2192,10 @@ class_edit(xo)
 {
   if (HAS_PERM(PERM_ALLBOARD | PERM_BM))
   {
-    short *chp;
+    int *chp;
     int chn;
 
-    chp = (short *) xo->xyz + xo->pos;
+    chp = (int *) xo->xyz + xo->pos;
     chn = *chp;
     if (chn >= 0)
     {
@@ -2262,12 +2262,12 @@ class_newbrd(xo)
 
   if (xo->key < CH_END)		/* 在分類群組裡面 */
   {
-    short *chx;
+    int *chx;
     char *img, *str;
     char xname[BNLEN + 1];
 
     img = class_img;
-    chx = (short *) img + (CH_END - xo->key);
+    chx = (int *) img + (CH_END - xo->key);
     str = img + *chx;
 
     str_ncpy(xname, str, sizeof(xname));
@@ -2294,19 +2294,19 @@ static int
 class_browse(xo)
   XO *xo;
 {
-  short *chp;
+  int *chp;
   int chn;
 
-  chp = (short *) xo->xyz + xo->pos;
+  chp = (int *) xo->xyz + xo->pos;
   chn = *chp;
   if (chn < 0)		/* 進入分類 */
   {
 #if 1
     /* smiler.070602: for熱門看板 */
-    short *chx;
+    int *chx;
     char *img, *str;
     img = class_img;
-    chx = (short *) img + (CH_END - chn);
+    chx = (int *) img + (CH_END - chn);
     str = img + *chx;
 
     if (!strncmp(str, "HOT/", 4))	/* "HOT/" 名稱可自定，若改名也要順便改後面的長度 4 */
@@ -2418,7 +2418,7 @@ static int
 class_addMF(xo)
   XO *xo;
 {
-  short *chp;
+  int *chp;
   int chn;
   MF mf;
   char fpath[64];
@@ -2426,7 +2426,7 @@ class_addMF(xo)
   if (!cuser.userlevel)
     return XO_NONE;
 
-  chp = (short *) xo->xyz + xo->pos;
+  chp = (int *) xo->xyz + xo->pos;
   chn = *chp;
 
   if (chn >= 0)		/* 一般看板 */
@@ -2453,11 +2453,11 @@ class_addMF(xo)
   }
   else			/* 分類群組 */
   {
-    short *chx;
+    int *chx;
     char *img, *str, *ptr;
 
     img = class_img;
-    chx = (short *) img + (CH_END - chn);
+    chx = (int *) img + (CH_END - chn);
     str = img + *chx;
 
     memset(&mf, 0, sizeof(MF));
@@ -2488,7 +2488,7 @@ MFclass_browse(name)
   char *name;
 {
   int chn, min_chn, len;
-  short *chx;
+  int *chx;
   char *img, cname[BNLEN + 2];
 
   min_chn = bshm->min_chn;
@@ -2499,7 +2499,7 @@ MFclass_browse(name)
 
   for (chn = CH_END - 2; chn >= min_chn; chn--)
   {
-    chx = (short *) img + (CH_END - chn);
+    chx = (int *) img + (CH_END - chn);
     if (!strncmp(img + *chx, cname, len))
     {
       if (XoClass(chn))
@@ -2524,7 +2524,7 @@ XoAuthor(xo)
   XO *xo;
 {
   int chn, len, max, tag, value;
-  short *chp, *chead, *ctail;
+  int *chp, *chead, *ctail;
   BRD *brd;
   char key[30], author[IDLEN + 1];
   XO xo_a, *xoTmp;
@@ -2540,12 +2540,12 @@ XoAuthor(xo)
   str_lower(author, author);
   len = strlen(author);
 
-  chead = (short *) xo->xyz;
+  chead = (int *) xo->xyz;
   max = xo->max;
   ctail = chead + max;
 
   tag = 0;
-  chp = (short *) malloc(max * sizeof(short));
+  chp = (int *) malloc(max * sizeof(int));
   brd = bshm->bcache;
 
   do
@@ -2686,7 +2686,7 @@ Class2()
 {
 #ifdef	MEICHU_WIN
   int chn, min_chn;
-  short *chx;
+  int *chx;
   char *img, *str;
   const char *name = "NthuMeichu/";
   
@@ -2697,7 +2697,7 @@ Class2()
   
   for (chn = CH_END - 2; chn >= min_chn; chn--)
   {
-    chx = (short *) img + (CH_END - chn);
+    chx = (int *) img + (CH_END - chn);
     if (!strncmp(img + *chx, name, strlen(name)))
     {
       more("brd/MeichuWin/note", NULL);
